feat: add the ability to sign in via email
When you sign in with correct credentials, you will get an API key.
This key is a json web token that contains user id and meta information
like when the token was issued and its expiration time.
This token is necessary for making sensitive API calls that require
authentication. In order for the server to see this API key you need
to attach this token to the HTTP Authorization header:
Authorization: Bearer YOUR_API_KEY
If the credentials you specified when signing in are wrong, you will
get the 401 error.