feat: add a default avatar
Now every user has the "avatar_url" column in the database.
Technically this column isn't allowed to be NULL, but if you don't
specify the avatar when using the signUpUser method in the auth
controller, this method still has a default value for this column.
This default value is specified in process.env.DEFAULT_AVATAR_URL.