feat: add the ability to sign up/log in with Google
Flow:
- (/auth/google-oauth-link) - get a link to the google authorization
server. The user will use this link to go to the consent page.
- if the user consents, they will be redirected to the redirect URI
(/auth/google-oauth-callback). The controller that handles this route
will receive an authorization code and it will exchange it for the
access token and id token. After that, the controller will retrieve
user info using the id token and check if the user is signed up and if
so, the API key will be generated. If the user doesn't exist, they will
be signed up and the API KEY will be generated.

One thing to take into account: if the user signs up with the Google
account, their account is automatically verified (email confirmation
is not required).