feat: add the ability to sign up/log in with Facebook
Now the oAuth flow has changed a bit:
- you request a link with these routes:
/auth/oauth-link/facebook
or
/auth/oauth-link/google
- the user goes to the consent page
- if the user consents, they will be redirected to the callback
(process.env.OAUTH_REDIRECT_URI). This callback will receive an
authorization code. This code will then be exchanged for the access
token (Facebook) or id token (Google). The goal of these is to get
first name, last name, email and avatar URL. If the email doesn't
exist in the database, a new account will be created. No matter if
the account exists or not, the goal of this callback is to return the
API key that can be used for further requests that require
authentication.