refactor: remove the Controller interface
Previously, a variable called "controllers", which implemented the
Controller interface (interfaces/Controller.ts) was exported from all
the controllers (for now it's only controllers/auth.ts). This variable
stored all the functions that would be executed when the user hits a
needed route. However, this way has proven to be wrong, because when
you specify some controller in the route, the editor doesn't provide
any suggestions. So isntead it would be better to export controller
functions using "export const controllerName: RequestHandler = ..."