fix(interface VerifiedUserInfo): replace 'userId' with 'id'
If the client attaches API_KEY in the request, this API key is then
verified. If it's valid, the data it holds (userId, expiration time,
issuance time) is put in the req.user object. This req.user adheres to
the VerifiedUserInfo interface. However, there is a mistake in there:
there should be a field "id" of type number instead of "userId".