feat: create/pay for orders with LiqPay
This adds the ability to create orders with this route:

POST /user/order/
It requires these fields:
- phoneNumber
- deliveryMethodName (based on DB table deliveryMethods)
- paymentMethodName (based on table orderPaymentMethods)
- cityName (table cities)
- postOffice (it's validated using NovaPoshta API or Ukrposhta API)
As the result, it returns an id of the newly created order.

If the paymentMethodName is 'Оплатити зараз', the user will have to pay
right after the order is created.
For the user to be able to pay, the client (React, mobile app etc) will
have to get the "data" and "signature". These fields are necessary to
redirect the user to the LiqPay payment page.
You can get these by hitting this route:

GET /user/order/:orderId/payment-data
It requires only orderId in the URL
This will return data and signature. When you get these, create a form.
An example of this form:
https://www.liqpay.ua/en/documentation/data_signature
(check the bottom of the page)

After the user pays for the order (or if the payment method is other
than 'Оплатити зараз') a notification will be sent with Telegram API
that describes the order. It will contain user info (username,
phone number to call the user and ask whether they confirm this order
along with information about the products they ordered as well as
delivery information).