feat!: add the ability to add products with GraphQL
This adds a new mutations called addProduct.
Example:
mutation {
    addProduct(
        title: "test",
        price: 123,
        previewURL: "http://test.com/image",
        quantityInStock: 6
    ) {
        id
    }
}

To add a product, an API_KEY should be provided, and the user who holds
this key must be an admin and have an activated account, otherwise the
request will fail.

This mutation required changing the payload of the API key (JWT key).
Now in addition to userId it also contains whether the user is an admin
and whether the user's account is activated.
With that said, a problem will arise in the future when a client
(React app for example) will need to refresh the API key when the
account gets activated, otherwise the key will be inconsistent, and the
server will think that this user's account is still not
activated.

Unit tests are also provided to this feature. They are written with
Vitest package.