refactor: remove port mapping for the API service
Previously, to access the API you'd need to go to http://localhost:3000
Now we're using nginx for traffic routing, so we don't need the API to
work standalone. You can now access it with http://localhost:80/api

This change is necessary so that there is no confusion in the future.