ci!: add GitHub Actions config
A new integrate.yml file was added to configure GitHub Actions.
When a pull request gets created, GitHub will automatically run tests.

(!) There was a bug with vitest when I tried to run:
> vitest run --watch false
A simple fix was to upgrade Node.js from v14.16.1 to v14.18
More about this error here:
https://github.com/vitest-dev/vitest/issues/670