test: generate a unique user before testing
This changes the way users are generated before testing components.
In the past, before any test is executed, a user with the same email
address was generated because this user would be wiped out anyway
because every test was assumed to be run in an isolated transaction.
However, some tests would require the user to be visible to other
transactions. For example, if you would need to update a product
using GraphQL, here are the steps:
 - Create a new user (isolated transaction, nobody else would see this
user except the test suite).
 - Create a product (it's part of that isolated transaction as well)
 - Update a product. In this case this action is done with the API
server (POST /graphql), meaning it can't see the changes to the DB
made by the two previous steps, and thus updating a product would
result in an error.
To solve this problem, we can prevent the first and the second step
from using isolated transactions, but in this case we would create
a new user over and over again (and the DB has mechanisms to prevent
adding the same users), and this commit aims to prevent that error.
Now every time a user is generated with the test suite, it would be
completely different from what user was generated the last time.
Now you may use isolated transactions in a test suite, or you may not,
it's up to you.