feat: add the ability to update products with GraphQL
This adds a new mutation called updateProduct.
Example:
mutation {
    updateProduct(
        id: 100,
        title: "updated title",
        price: 1234,
        previewURL: "http://mysite.com/new-preview-image.png",
        quantityInStock: 10
    ) {
        id
        title
        price
        previewURL
        isAvailable
        isRunningOut
    }
}
Strictly speaking, only 'id' field is required to specify when editing
a product, BUT you will get an error anyway because the action of
determining whether there are other arguments other than 'id' is
done within a resolver.

As with the addProduct mutation, this updateProduct mutation requires
a user to be authenticated, activated and to be an admin as well.
So only admins who have activated accounts can execute this mutation.

(!) It is also important to note that the updateProduct mutation
accepts this data:
{
    id: number;
    title: string;
    price: number;
    previewURL: string;
    quantityInStock: number;
}
BUT it returns this data:
{
    id: number;
    title: string;
    price: number;
    previewURL: string;
    isAvailable: boolean;
    isRunningOut: boolean;
}