fix: add product helper functions to avoid duplication
This commit can be described both as a fix and a refactor.
When GraphQL resolvers received 'quantityInStock' from a user,
they tried to determine whether the product is available and whether
the product is running out by running the same set of instructions.
This can lead to inconsistency and code repetition, and that was
exactly the case. Some resolvers considered a product available
if its quantity is 2 or more, and some - 1 or more, thus leading to an
inconsistency. That's why two separate helper functions were created
(called isProductAvailable and isProductRunningOut).
Now if you want these criteria to change (for example now the product
is considered to be running out if there are 10 samples of it left),
instead of changing every resolver you can just change these two files
and the changes will be applied to all the resolvers.