feat: add the ability to delete products with GraphQL
This adds a new mutation called deleteProduct.
Example:
mutation {
    deleteProduct(id: 100) {
        id
    }
}
As you can see, this mutation returns an id of the deleted product.

(!) If the specified product id does not exist, this mutation will
return an error.
(!) As with the previous mutations (addProduct, updateProduct),
only admins who have activated accounts can execute this mutation.
(in other words, a valid API_KEY is required).