chore: disable the automatic deployment process
This commit disables the automatic deployment process for the
application using GitHub Actions. The process was consuming AWS
resources and could result in unnecessary charges in the future.
The deploy job in the GitHub Actions workflow has been simply
commented out. No functional changes were made to the application.
It's also important to point out that the CI/CD process that has been
commented out consisted of these steps:
- Set up the project
- Run tests on the API server
- If they pass, deploy the application to Google Cloud
This is important because now this project uses React.js as well.
So if this deployment process works again at the some point in the
future, the necessary changes will need to be made to ensure that
automatic tests on the React project are run as well.