feat: implement signup flow using Formik
This adds the ability for users to sign up to the application by
going to http://localhost/sign-up.

I chose to use Formik because my initial try to create the signup
flow had lots of repetitive code, I had to manage state for values,
touched state and so on. And all of this for all the inputs.
Formik abstracts all of that and gives you a lot of utility
props/methods which allow you to manage forms much easily.

The inputs of the form have a custom onChange/onBlur handler, because
using Formik's ones aren't very flexible. I found that using Formik's
native event handlers resulted in validating the entire form on every
keystroke, and that was horrible especially for the API call that
checks email availability. To overcome this, I created custom handlers
that update values and the touched state and then validate only one
field instead of the all the fields, which Formik's native handlers do.

Every input of the form works with the touched state. The validation
errors are only shown when the touched state is set to true.

Since the /api/is-email-available route has a rate limit of 1 second,
meaning you can send a request only once every 1 second, I implemented
the debounce function to limit the number of API calls. One thing to
note about this debounce function: the function that you provide there
must return a promise. Another thing is that this function works in
similar fashion to the debounce function of lodash but with leadingEdge
set to "true" (the debounced function will be invoked immediately when
it is first called) This allows a user to make an API call immediately
if this is the first time calling the function or if the cooldown
period has passed. The reason why I didn't use the same function of
some library, for example lodash, is that it didn't provide any means
to find out whether the function started executing, which is needed
in order to show the loading spinner until the API call is finished.