feat: add GET /is-email-available route
This commit introduces a route that returns whether an email is taken
or not by checking the database. This will be useful for signing up.
This route expects an email as a query parameter, for example:
GET /is-email-available?email=test@test.com
It returns a JSON object containing a boolean field 'isEmailAvailable'
indicating the availability of the email.

Before checking the database, the server validates the email. If the
email is not a correct format, the server will return a response with
a 422 status code.

Exposing this route without any protection is quite dangerous because
an attacker could simply bombard the API server with different requests
to find out which emails are taken and which are not. To mitigate this
risk, a rate limiting system was implemented with nginx. This makes
sure that only 1 request per second can be sent per IP address.
This is not a foolproof solution but it does provide a significant
improvement in the security of the application.