feat: add a success message to the signup form
If a user successfully signs up and the server returns a 201 status
code, the user will see a success message that says that a confirmation
email has been sent to the specified email. They would need to follow
the link to activate their account (right now account activation is
only implemented on the server side).