feat: client-side account activation implementation
This commit implements the client-side account activation feature.
Upon sign up, a confirmation email containing a link is sent to the
user's email address. Upon clicking the link, the user is redirected
to a React.js page that sends a request to the API to activate their
account.

With this change, users can now activate their account directly from
the client-side, improving the overall sign-up experience.