refactor: change status code from 403 to 422 in OAuth 2.0 callback route
When the user wants to sign in/up with Google/Facebook, they are
redirected to the authorization server of that service. When the user
gives their consent, they are redirected to the "callback" route.
This route is managed by OAuthCallback controller method.
This route accepts the "state" parameter, and this parameter should
indicate that the request was made by Google/Facebook and nobody else.
After verifying this parameter, if this "state" parameter is incorrect,
the controller returns the 403 error with a message
"Invalid 'state' parameter". However, this route should return the 422
status code instead of 403, because it's more about validation than
authorization.