fix: change status code from 422 to 500 in OAuth 2.0 callback route
The OAuth 2.0 callback route accepts the "state" query parameter,
which is used to ensure that the request was made by the intended
authorization server (Google/Facebook).

After the "state" parameter is validated, the OAuth 2.0 callback route
tries to determine the name of the resource server (Google/Facebook) by
using this "state". If for some reason the server controller receives
a result that's different from "google" or "facebook", an error is
thrown with the status code of 422.

However, this status code indicates that the request is malformed,
which is not appropriate in this case. If the route receives anything
other than "google" or "facebook", it would be the server's fault, and
the 500 status code is more appropriate. This is because the "state"
parameter has already been validated, and if the execution flow reaches
the point of determining the name of the resource server, it means that
the "state" parameter is semantically correct.