fix: omit 'id' in the 'addProduct' GraphQL resolver
The addProduct resolver accepts 'args' as a parameter.
Previously, it was specified that this argument is of type DBProduct,
which accepts these fields:

id: number;
title: string;
price: number;
previewURL: string;
quantityInStock: number;

However, this is not entirely correct, because 'id' is not supposed
to be specified as an argument in the addProduct resolver. This is why
a type was replace to one that omits this 'id' property.

This is not a problem per se, but it can mislead future developers and
cause errors in the future.