feat!: introduce access/refresh tokens to authenticate a user
This commit changes the way authentication is handled in the REST API.

Previously, a simple API_KEY was used to authenticate a user. This
API_KEY is a JWT that holds the userId and other meta information.
However, this method is not very viable, because the server returned it
in the response body, which means that clients would need to store it
either in localStorage or in sessionStorage, and this, in turn, would
make users vulnerable to XSS attacks.

Another disadvantage of this method is that it's really difficult to
revoke tokens without implementing sessions. As we know, a key advantage
of using JWT is not having to reach to the database, because all the
information is stored in the token itself. But what if a user gets
hacked or if they want to terminate all their sessions from other
devices? It's really difficult to do so by using JWTs that are not
stored anywhere.

As a side note, we could use session-based authencation (a user
receives a session id that points to some data that's stored on the
server). However, reaching to the database to verify the session id
every time the user sends a request isn't a good idea from the
performance point of view. This is why access tokens and refresh tokens
were introduced to get the best of both worlds (fast accessToken
validation while being able to invalidate sessions (with some buts)).

Let's first clarify the terminology:
- An access token is attached to every sensitive requests that require
a user to be authenticated. This access token is a JWT. Access tokens
are typically short-lived. Their typical lifespan is 10-30 minutes, it
depends on the application. They are typically stored in localStorage
or, even better, in memory.
- A refresh token is used to request new access tokens. They are
typically long-lived, for example a month, a year, or for an indefinite
period of time or until it's revoked. They are typically stored as a
cookie (but it's important to make sure this cookie isn't vulnerable to
XSS and CSRF attacks!).

Let's see how this flow works:
- A user logs in successfully to the React.js application. The server
sends an access token and a refresh token. The access token would be in
the response body, but the refreshToken is sent as a cookie that has
the following options:
    1. SameSite = "strict" (the browser won't send the cookie to any
    3rd party website)
    2. HttpOnly (the cookie is not accessible from JavaScript)
    3. path = "/api/auth/refresh" (it scopes the cookie to be included
    only if a request is being sent to the "/api/auth/refresh" route).
    4. MaxAge = 1 month (it's configurable in the .env file).
    5. When this website has its own SSL certificate, this cookie would
    need to have the "secure" option as well. Right now it's not the
    case.
- The client would store the access token in memory (in our case by
using Redux). The reason why I chose the in-memory option is because
storing access tokens in localStorage would make them vulnerable to XSS
attacks. Storing access tokens in memory is probably not 100% safe, but
it would require much more sophisticated approaches to steal the access
token. However, this approach has its downsides, one of which is that
the access token would be lost when the page reloads. But since we have
a refresh token as a cookie that's NOT accessible from JavaScript, this
cookie would be used to retrieve a new access token by making the
request to /api/auth/refresh. The refreshToken cookie would be sent
there as well. After receiving the access token we simply put it in the
Redux store again. Now we can make sensitive requests by attaching this
access token to the 'Authorization' header.

Let's summarize why access/refresh approach was used by comparing the
tradeoffs of other authentication methods:
1. Session-based authentication:
 + Easy to understand.
 + Easy to implement.
 + Easy session revokation.
 - Slow session id verification! We would need to reach to the database
on every request.
2. Using an API_KEY, which is a JWT token:
 + Easy to understand.
 + Easy to implement.
 + Fast token verification.
 - Difficult revokation. That would basically require us to reimplement
the session-based authentication.
3. Access/refresh tokens:
 + Fast token verification, BUT only with access tokens because they are
JWTs that are NOT stored in any DB.
 + Easy revokation, BUT we can invalidate only refresh tokens, because
they are stored in the database, but we can't revoke access tokens, so
a user would still have some time to use of our application, but in this
application it's fine. This is why it's important to make access tokens
short-lived. Another reason to reduce the lifespan of access tokens is
XSS and CSRF attacks. If a maliciou user does acquire an access token
of another user, they must have as little time as possible to use it.
 - More difficult to understand.
 - More difficult to implement.

As we can see, none of the approaches are perfect, but considering all
the tradeoffs, using access/refresh tokens seems the best option in this
application.