fix: use a single client when using transactions
Previously, the REST API server used the database pool to create
transactions, but this can lead to issues because PostgreSQL isolates
transactions to individual clients. To fix this, this commit changes
the way transactions are handled so that a single client is used per
transaction.

More about this issue can be found here:
https://node-postgres.com/features/transactions

This change also required to refactor some model methods:
 - authModel.signUpUser
 - authModel.addActivationTokenToDB
 - orderModel.createOrder
 - orderModel.transferCartProductsToOrderProducts
 - cartModel.cleanCart
These methods now accept an optional 'dbClient' argument that must be
of type 'PoolClient'. This client is derived by using
"await dbPool.connect()". If this argument is not specified, dbPool
will be used instead.

Another implication of this change is changing the way transactions are
managed. Previously, this was done by using 'helperModel' that has
three methods: 'beginTransaction', 'commitTransaction',
'rollbackTransaction'. This is still the case, but now these methods
accept a mandatory argument 'dbClient' of type PoolClient.