fix!: enforce prior account activation before signing in
Previously, when the user tried to sign in to their account, they could
do it regardless of the account activation status. This commit changes
this behavior. Now if the user is trying to log in, their account must
be activated, otherwise a 403 error will be returned.

Because of this change, it's no longer necessary to hold the
'isActivated' field in the JWT API_KEY. Firstly, it's simply redundant,
because if the user signed in, it already implies that their account is
activated. Secondly, it would be difficult to keep this field
consistent, because it would require to invalidate all the previous
tokens, which is pretty difficult considering that the server just
returns an API_KEY, which is a simple JWT that is not stored in any DB.