fix: restrict product changes to activated admin users in GraphQL requests
Previously, unactivated admin users could make GraphQL requests to
change products in this application, which was a major security
vulnerability. This commit fixes this issue. Now, any attempt to
change, delete or add products will be restricted to authorized admin
users only. This change helps to prevent unauthorized access to
sensitive data and improves the overall security of the application.