feat: manually start sending a request with the 'useFetch' hook
This commit changes the way the useFetch hook works. Instead of sending
the request automatically once the hook was declared in some component,
now it's necessary to import the "sendRequest" function from this
custom hook and then call this function whenever there's a need to send
the request.

Despite sounding like a limitation, this actually makes the hook more
versatile, because in most cases you want a request to be sent not
immediately, but whenever something important happens.