refactor: use 'react-google-recaptcha' for captcha logic
This commit changes the way ReCAPTCHA is handled. Instead of writing
the <ReCAPTCHA /> component from scratch, now this custom component
uses the ReCAPTCHA component from the 'react-google-recaptcha' package.
Previously, to connect the captcha it was necessary to include a script
in index.html (not in the component because otherwise it wouldn't work).
This new change makes the recaptcha more modular.

This change would prevent a bug where the captcha isn't loaded when
trying to create protected routes.