feat: add <ErrorMessage /> component to show text errors
This component is used to show error messages in form of a paragraph.
It was added because there would be a lot of code duplication around
showing errors otherwise. For example, the sign up page shows error
messages when a user entered invalid text to inputs.
In the future, when there will be a login page, error messages will
need to be shown as well, and this component will be used to show them.
You won't have to write error messages from scratch.

This component accepts three props, two of which are optional:
    - message (mandatory, string): error message that will be shown to
the user;
    - centered (optional, boolean, default - false): whether the
message is centered or not;
    - addBackground (optional, boolean, default - false): whether to
add a red background to the text.

Examples of using this component
<ErrorMessage centered={true} addBackground={true} message="Error!" />
<ErrorMessage centered={true} message="Something went wrong" />
<ErrorMessage addBackground={true} message="Invalid credentials" />

There is already a similar component that's called ErrorNotification
and it serves the same purpose: show an error to the user, but
ErrorNotification does it differently. Instead of returning a paragraph,
this component is always shown on the top-left side of the screen and
it's automatically closed after some short period of time (typically 5
or 10 seconds) or a user can close it manually by simply clicking on it.
The <ErrorMessage /> component, on the contrary, cannot be closed at
all. It's just a text that's typically shown in some form due to
validation errors.