feat!: implement client-side sign-in process
- Users can sign in on the /sign-in route using email or phone number
and password.
- Form validation checks for correct format of phone numbers and
non-empty password.
- The phone number must be like these examples: +380123456789 or
+380-12-345-67-89.
- Captcha challenge was added to enhance security.
- Server response handles account activation status and returns refresh
and access tokens on successful login.
- Access token is stored in Redux, refresh token is stored as
a cookie with HttpOnly, SameSite, and Path=/api/auth/refresh flags for
XSS and CSRF protection.

This commit also makes the /sign-in route protected, meaning a user can
visit it ONLY when they are NOT authenticated, otherwise they will be
redirected to /. This was made possible by introducing the
<EnsureStatus /> component. This component accepts these props:
- auth: boolean. If 'auth' is set to true, the user must be
authenticated to access the route, otherwise the user must be
unauthenticated
- children: React.ReactNode. This specifies the concrete component that
will be rendered in a route if the auth condition is met. For example:
<Route
    path="/sign-in"
    element={
        // the user must be unauthenticated to access this route
        <EnsureStatus auth={false}>
            <SignIn />
        </EnsureStatus>
    }
/>

To make the /sign-in route protected, the <App /> component was
modified. When the React application loads up, the client makes a
request to the /api/auth/refresh route. The browser will automatically
attach the refresh token if it's set as a cookie. The server will
verify this refresh token and if it's correct it will return an access
token. The client will store this access token in the Redux store.