refactor: create <PasswordTips /> component for password requirements
This commit encapsulates password tips by creating a new component that
will hold password requirements.
Previously, password requirements were included in the <SignUpForm />
component, but this won't allow us to reuse the password requirements
component in case we need it in the future. For example, this change
will be useful when we create the "forgot password" feature. We will
need this <PasswordTips /> component there.