feat(client): suggest a user to activate their account on sign in
When a user enters valid credentials and their account is not
activated, they won't be allowed to sign in. Instead, an error message
will be shown to them and they will be asked whether they want to send
a new activation link to their email. If they choose to do so, the link
will be sent to their email and they will have to follow it to activate
their account.