feat: use Redis for storing activation tokens
This commit replaces the previous implementation of account activation
logic using PostgreSQL with a new implementation using Redis.

When a user signs up, an activation token is generated and stored in a
Redis database. This token is used to activate the user's account
within a one-hour window.

Using Redis for activation token storage provides several benefits
over the previous implementation, including faster access times,
automatic expiration of activation tokens, and reducing the burden on
PostgreSQL.

Right now Redis is run locally by using docker-compose, but when the
application is ready to be deployed, it's possible that it will use a
production-ready Redis cluster, for example by using AWS ElastiCache.