fix: specify recaptcha size on page load
Previously, after adding the recaptcha logic on the client, the login
and signup pages would become unresponsive if the page width was less
than ~400px. This was because the recaptcha component didn't
automatically respond to the page width changes.

To fix this, I initially set the 'size' prop to 'compact' or 'normal'
depending on the page width. I tried changing the 'size' prop on page
resize, but the component didn't re-render, as documented in this issue:
https://github.com/dozoisch/react-google-recaptcha/issues/69
However, it's unlikely that users will change the page size, so setting
the 'size' prop initially is sufficient to solve this issue.