feat: add an endpoint to resend an activation link
This commit adds a new endpoint:

POST /api/auth/resend-activation-link
What it accepts (application/json):
    - login: string (email or phone number)
    - password: string
    - recaptchaToken: string
What it returns:
    - targetEmail: string (which email the link was sent to)
Possible errors:
    - 422 (if the specified body fields are empty)
    - 409 (conflict - the account is already activated)
    - 500 (an unexpected error occurred - server fault)
(!) This route implements rate limiting. A single IP address can only
send 1 request per second to prevent brute force attacks.

If a user tries to sign in with an account that's not activated, they
will be prompted to send a new activation link to their email.
Login and password will already be there, because the user has just
tried to sign in, so all that's needed from a user is to just solve the
recaptcha challenge.