fix: export a timestamp from useFetch for re-renders
When components that use this hook send a request and it fails, it
works just fine. The state is changing, and eveything works as
expected. But when the same request is made again by using
"sendRequest()" and the result of this request is the same, components
don't get updated. Because the state is the same, React.js doesn't
cause components to re-render.

To solve this issue, the useFetch hook now exports a unique state
called "requestTimestamp". It will change to the current timestamp
right after the request finishes executing. It doesn't matter whether
the request succeeded or failed - requestTimestamp will change anyway.
This can be used in the dependency array of useEffect to execute the
function even if the result of the second request is the same as the
first one.