refactor: move error handling logic from useReducer to Redux
This commit changes the way errors are handled in the React app.
Error handling logic was movied from useReducer to Redux for better
code organization, reduction of code duplication and scalability.

Now, the <App /> component listens to changes in the error slice and
displays the <ErrorNotification /> component when errors occur. This
change increases the maintainability of the codebase and makes it
easier to add new features in the future.