refactor: migrate /sign-up to RTK Query
This commit changes the way requests are handled in the /sign-up route
of the React application.
Instead of using a custom hook for making API requests, this commit
starts to migrate making API requests from custom hooks to using RTKQ.

A couple of benefits of using RTK Query:
- RTK Query automatically handles caching and deduplication of API
requests, which can greatly improve performance and reduce unnecessary
network traffic.
- By using a widely adopted library like RTK Query, it's possible to
reduce the amount of custom code needed to write and maintain, and
benefit from ongoing community support and development.