fix: reset activation form when closing a modal window
When a user signed in with valid credentials but their account is not
activated, they will be asked if they want to send a new activation
link to their email. If they want to do so, they will need to solve the
recaptcha challenge, and then they will be able to send the request to
resend the activation link.

However, if the user solved the captcha challenge and then immediately
closed the modal window, and then they would open this modal window
again, they would still be able to send the form (even though visually
the recaptcha has been reset). This behavior can lead to bugs.
This commit fixes this. Now when the user closes the modal window with
the solved captcha, the form that contains the captcha will be reset
and the user won't be able to send the request without completing the
recaptcha challenge again.