fix(client): remove flickering on account activation
When a user opens a link from their email to activate their account,
they will be redirected to /auth/activate-account/:activationToken.
This React.js page would almost immediately start a request to the
server to activate the user's account. However, prior to this commit
the loading spinner logic was implemented improperly for this page.
When a user enters the page, they could see the error message for a
very brief moment, then the loading spinner would show up, and then the
success message if everything went smoothly.

This commit fixes the loading logic and now there is no "flickering"
when the page is shown to the user.