feat(API): implement POST /api/auth/send-reset-token
This route is used to request a link to reset a user's password.
It takes these values as the request body:
- recaptchaToken
- email
After this, the reset token will be generated and a special link will
be sent to the user's email.
The link that was sent will look like this:
http://{domain}/auth/reset-password/:resetToken
By default, the link expires in 1 hour.

If the data that was sent is incorrect (either of these cases):
- the recaptcha token has expired
- the email is incorrrect
- there is no user with the corresponding email
If any of these cases are true, the server will return the status code
of 422.

If everything goes as expected, the server will send a 200 status code.