feat: add a page to request a link to reset the password
If a user forgets their password, they can follow the link that's on
the /sign-in page. After following the link they will be redirected to
/forgot-password. They will be asked to enter their email and solve the
captcha challenge. After sending the form they will either get a
message that tells them that everything went fine and they should check
their email for the link, or they will get an error. This error can be
that the email doesn't exist (brute force protection is implemented on
the API level by enforcing the recapcha validation), or a generic error
like "Something went wrong" is the server is offline.